タスク11の修正

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import jp.co.seattle.library.service.BooksService;
+import jp.co.seattle.library.service.RentBooksService;
 
 /**
  * 削除コントローラー
@@ -22,6 +23,9 @@ public class DeleteBookController {
     final static Logger logger = LoggerFactory.getLogger(DeleteBookController.class);
     @Autowired
     private BooksService booksService;
+    
+    @Autowired
+    private RentBooksService rentBooksService;
     /**
      * 対象書籍を削除する
      *
@@ -37,7 +41,16 @@ public class DeleteBookController {
             @RequestParam("bookId") Integer bookId,
             Model model) {
         logger.info("Welcome delete! The client locale is {}.", locale);
-        booksService.deleteBook(bookId);
+        int rentedBookId = rentBooksService.getRentBookInfo(bookId);
+        
+        if(rentedBookId == 0) {
+        	booksService.deleteBook(bookId);
+        } else {
+        	model.addAttribute("error", "貸出中のため削除できません。");
+        	model.addAttribute("bookDetailsInfo", booksService.getBookInfo(bookId));
+        	return "details";
+        }
+        
         model.addAttribute("bookList", booksService.getBookList());
         return "home";
